backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydub==0.25.1  # Requires ffmpeg
openai-whisper==20231117
sentence-transformers
huggingface-hub>=0.16.0
SQLAlchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
pydantic-settings==2.0.3
pytest==7.4.3
httpx==0.25.1
numpy<2.0.0  # Pin numpy to 1.x version
spacy==3.7.2
opensearch-py==2.3.1
boto3==1.34.0
requests-aws4auth==1.2.3

# System requirements:
# ffmpeg - Install via:
# macOS: brew install ffmpeg
# Ubuntu: apt-get install ffmpeg
# Windows: choco install ffmpeg 